housewarming_backend/requirements.txt
Django==6.0.4
python-dotenv==1.1.1
